+++ jgclark.DashboardReact/src/dashboardHelpers.js
@@ -6,18 +6,12 @@
 
 // import moment from 'moment/min/moment-with-locales'
 import pluginJson from '../plugin.json'
+import { removeDateTagsAndToday, getAPIDateStrFromDisplayDateStr, includesScheduledFutureDate } from '../../helpers/dateTime'
 import type { TSectionItem, TParagraphForDashboard } from './types'
 // import { showDashboard } from './HTMLGeneratorGrid'
 import { getSettingFromAnotherPlugin } from '@helpers/NPConfiguration'
 import { clo, JSP, logDebug, logError, logInfo, logWarn, timer } from '@helpers/dev'
-import {
-  getAPIDateStrFromDisplayDateStr,
-  includesScheduledFutureDate,
-} from '@helpers/dateTime'
-import {
-  createRunPluginCallbackUrl,
-  displayTitle,
-} from "@helpers/general"
+import { createRunPluginCallbackUrl, displayTitle } from '@helpers/general'
 import { filterOutParasInExcludeFolders } from '@helpers/note'
 import { getReferencedParagraphs } from '@helpers/NPnote'
 import {
@@ -30,15 +24,8 @@ import {
   smartPrependPara,
 } from '@helpers/paragraph'
 import { findParaFromStringAndFilename } from '@helpers/NPParagraph'
-import {
-  RE_ARROW_DATES_G,
-  RE_SCHEDULED_DATES_G,
-} from '@helpers/regex'
-import {
-  addPriorityToParagraphs,
-  getNumericPriorityFromPara,
-  sortListBy,
-} from '@helpers/sorting'
+import { RE_ARROW_DATES_G, RE_SCHEDULED_DATES_G } from '@helpers/regex'
+import { addPriorityToParagraphs, getNumericPriorityFromPara, sortListBy } from '@helpers/sorting'
 import { eliminateDuplicateSyncedParagraphs } from '@helpers/syncedCopies'
 import {
   getTimeBlockString,
@@ -46,15 +33,8 @@ import {
   // isTimeBlockPara,
   RE_TIMEBLOCK_APP,
 } from '@helpers/timeblocks'
-import {
-  displayTitleWithRelDate,
-  showMessage
-} from '@helpers/userInput'
-import {
-  isOpen, isOpenTask, isOpenNotScheduled, isOpenTaskNotScheduled,
-  removeDuplicates
-} from '@helpers/utils'
-import { removeDateTagsAndToday } from '../../helpers/dateTime'
+import { displayTitleWithRelDate, showMessage } from '@helpers/userInput'
+import { isOpen, isOpenTask, isOpenNotScheduled, isOpenTaskNotScheduled, removeDuplicates } from '@helpers/utils'
 
 //-----------------------------------------------------------------
 
@@ -110,9 +90,7 @@ export async function getSettings(): Promise<any> {
     const config: dashboardConfigType = await DataStore.loadJSON(`../${pluginID}/settings.json`)
 
     if (config == null || Object.keys(config).length === 0) {
-      throw new Error(
-        `Cannot find settings for the '${pluginID}' plugin. Please make sure you have installed it from the Plugin Preferences pane.`,
-      )
+      throw new Error(`Cannot find settings for the '${pluginID}' plugin. Please make sure you have installed it from the Plugin Preferences pane.`)
     }
     // clo(config, `settings`)
     // Set special pref to avoid async promises in decideWhetherToUpdateDashboard()
@@ -128,7 +106,7 @@ export async function getSettings(): Promise<any> {
     logDebug(pluginJson, `filter? -> ${String(DataStore.preference('Dashboard-filterPriorityItems'))}`)
 
     // Extend settings with a couple of values, as when we want to use this DataStore isn't available etc.
-    config.timeblockMustContainString = DataStore.preference("timeblockTextMustContainString") ?? ''
+    config.timeblockMustContainString = DataStore.preference('timeblockTextMustContainString') ?? ''
     config.filterPriorityItems = DataStore.preference('Dashboard-filterPriorityItems')
 
     return config
@@ -143,7 +121,7 @@ export async function getSettings(): Promise<any> {
 
 /**
  * Return an optimised set of fields based on each paragraph (plus filename + computed priority + title - many)
- * @param {Array<TParagraph>} origParas 
+ * @param {Array<TParagraph>} origParas
  * @returns {Array<TParagraphForDashboard>} dashboardParas
  */
 export function makeDashboardParas(origParas: Array<TParagraph>): Array<TParagraphForDashboard> {
@@ -163,8 +141,7 @@ export function makeDashboardParas(origParas: Array<TParagraph>): Array<TParagra
       }
     })
     return dashboardParas
-  }
-  catch (error) {
+  } catch (error) {
     logError('makeDashboardParas', error.message)
     return []
   }
@@ -186,7 +163,9 @@ export function makeDashboardParas(origParas: Array<TParagraph>): Array<TParagra
  * @returns {[Array<TParagraph>, Array<TParagraph>]} see description above
  */
 export function getOpenItemParasForCurrentTimePeriod(
-  timePeriodName: string, timePeriodNote: TNote, config: dashboardConfigType
+  timePeriodName: string,
+  timePeriodNote: TNote,
+  config: dashboardConfigType,
 ): [Array<TParagraphForDashboard>, Array<TParagraphForDashboard>] {
   try {
     let parasToUse: $ReadOnlyArray<TParagraph>
@@ -195,10 +174,13 @@ export function getOpenItemParasForCurrentTimePeriod(
     // Get paras from calendar note
     // Note: this takes 100-110ms for me
     const startTime = new Date() // for timing only
-    if (Editor && (Editor?.note?.filename === timePeriodNote.filename)) {
+    if (Editor && Editor?.note?.filename === timePeriodNote.filename) {
       // If note of interest is open in editor, then use latest version available, as the DataStore is probably stale.
       parasToUse = Editor.paragraphs
-      logDebug('getOpenItemParasForCurrent...', `Using EDITOR (${Editor.filename}) for the current time period: ${timePeriodName} which has ${String(Editor.paragraphs.length)} paras (after ${timer(startTime)})`)
+      logDebug(
+        'getOpenItemParasForCurrent...',
+        `Using EDITOR (${Editor.filename}) for the current time period: ${timePeriodName} which has ${String(Editor.paragraphs.length)} paras (after ${timer(startTime)})`,
+      )
     } else {
       // read note from DataStore in the usual way
       parasToUse = timePeriodNote.paragraphs
@@ -213,7 +195,7 @@ export function getOpenItemParasForCurrentTimePeriod(
     // Need to filter out non-open task types for following function, and any scheduled tasks (with a >date) and any blank tasks.
     // Now also allow to ignore checklist items.
     // Note: this operation is 100ms
-    let openParas = (config.ignoreChecklistItems)
+    let openParas = config.ignoreChecklistItems
       ? parasToUse.filter((p) => isOpenTaskNotScheduled(p) && p.content.trim() !== '')
       : parasToUse.filter((p) => isOpenNotScheduled(p) && p.content.trim() !== '')
     logDebug('getOpenItemParasForCurrent...', `After 'isOpenTaskNotScheduled + not blank' filter: ${openParas.length} paras (after ${timer(startTime)})`)
@@ -257,10 +239,10 @@ export function getOpenItemParasForCurrentTimePeriod(
     let refOpenParas: Array<TParagraph> = []
     if (timePeriodNote) {
       // Allow to ignore checklist items.
-      refOpenParas = (config.ignoreChecklistItems)
+      refOpenParas = config.ignoreChecklistItems
         ? getReferencedParagraphs(timePeriodNote, false).filter(isOpenTask)
-        // try make this a single filter
-        : getReferencedParagraphs(timePeriodNote, false).filter(isOpen)
+        : // try make this a single filter
+          getReferencedParagraphs(timePeriodNote, false).filter(isOpen)
     }
     logDebug('getOpenItemParasForCurrent...', `- got ${refOpenParas.length} open referenced after ${timer(startTime)}`)
 
@@ -318,8 +300,7 @@ function isTimeBlockLine(contentString: string, mustContainString: string = ''):
     }
     const res2 = contentString.match(RE_TIMEBLOCK_APP) ?? []
     return res2.length > 0
-  }
-  catch (err) {
+  } catch (err) {
     console.log(err)
     return false
   }
@@ -339,9 +320,8 @@ function isTimeBlockPara(para: TParagraph, mustContainStringArg: string = ''): b
     return false
   }
   const tbString = getTimeBlockString(para.content)
-  return (!isTermInURL(tbString, para.content))
+  return !isTermInURL(tbString, para.content)
 }
-
 
 // Display time blocks with .timeBlock style
 // Note: uses definition of time block syntax from plugin helpers, not directly from NP itself. So it may vary slightly.
@@ -365,15 +345,14 @@ function convertTimeBlockToHTML(input: string): string {
  */
 function parseAndSortDates(items: Array<TParagraph>): Array<ParsedTextDateRange> {
   const withDates = items
-    .map(item => ({
+    .map((item) => ({
       item,
-      date: Calendar.parseDateText(item.content)[0]?.start ?? null
+      date: Calendar.parseDateText(item.content)[0]?.start ?? null,
     })) // Map each item to an object including both the item and the parsed start date.
     .filter(({ date }) => date != null) // Filter out items without a valid start date.
 
   // Sort the intermediate structure by the start date and map back to the original items.
-  const sortedItems = withDates.sort((a, b) => a.date - b.date)
-    .map(({ item }) => item)
+  const sortedItems = withDates.sort((a, b) => a.date - b.date).map(({ item }) => item)
 
   return sortedItems
 }
@@ -422,7 +401,6 @@ export async function getRelevantOverdueTasks(config: dashboardConfigType, yeste
   }
 }
 
-
 /**
  * Make an HTML link showing displayStr, but with href onClick event to show open the 'item' in editor and select the given line content
  * @param {SectionItem} item's details, with raw
@@ -446,8 +424,7 @@ export function addNoteOpenLinkToString(item: SectionItem, displayStr: string): 
       // return `<a>${displayStr}</a>`
       return `${displayStr}`
     }
-  }
-  catch (error) {
+  } catch (error) {
     logError('addNoteOpenLinkToString', `${error.message} for input '${displayStr}'`)
     return '(error)'
   }
@@ -464,9 +441,10 @@ export function makeNoteTitleWithOpenActionFromFilename(item: SectionItem, noteT
   try {
     // logDebug('makeNoteTitleWithOpenActionFromFilename', `- making notelink with ${item.filename}, ${noteTitle}`)
     // Pass request back to plugin, as a single object
-    return `<a class="noteTitle sectionItem" {()=>onClickDashboardItem({itemID: '${item.ID}', type: 'showNoteInEditorFromFilename', encodedFilename: '${encodeURIComponent(item.para.filename)}', encodedContent: ''})}<i class="fa-regular fa-file-lines pad-right"></i> ${noteTitle}</a>`
-  }
-  catch (error) {
+    return `<a class="noteTitle sectionItem" {()=>onClickDashboardItem({itemID: '${item.ID}', type: 'showNoteInEditorFromFilename', encodedFilename: '${encodeURIComponent(
+      item.para.filename,
+    )}', encodedContent: ''})}<i class="fa-regular fa-file-lines pad-right"></i> ${noteTitle}</a>`
+  } catch (error) {
     logError('makeNoteTitleWithOpenActionFromFilename', `${error.message} for input '${noteTitle}'`)
     return '(error)'
   }
@@ -484,9 +462,10 @@ export function makeNoteTitleWithOpenActionFromTitle(noteTitle: string): string 
     // logDebug('makeNoteTitleWithOpenActionFromTitle', `- making notelink from ${noteTitle}`)
     // Pass request back to plugin
     // Note: not passing rawContent (param 4) as its not needed
-    return `<a class="noteTitle sectionItem" {()=>onClickDashboardItem({itemID:'fake', type:'showNoteInEditorFromTitle', encodedFilename:'${encodeURIComponent(noteTitle)}', encodedContent:''}}><i class="fa-regular fa-file-lines pad-right"></i> ${noteTitle}</a>`
-  }
-  catch (error) {
+    return `<a class="noteTitle sectionItem" {()=>onClickDashboardItem({itemID:'fake', type:'showNoteInEditorFromTitle', encodedFilename:'${encodeURIComponent(
+      noteTitle,
+    )}', encodedContent:''}}><i class="fa-regular fa-file-lines pad-right"></i> ${noteTitle}</a>`
+  } catch (error) {
     logError('makeNoteTitleWithOpenActionFromTitle', `${error.message} for input '${noteTitle}'`)
     return '(error)'
   }
@@ -504,9 +483,10 @@ export function makeNoteTitleWithOpenActionFromNPDateStr(NPDateStr: string, item
     const dateFilename = `${getAPIDateStrFromDisplayDateStr(NPDateStr)}.${DataStore.defaultFileExtension}`
     // logDebug('makeNoteTitleWithOpenActionFromNPDateStr', `- making notelink with ${NPDateStr} / ${dateFilename}`)
     // Pass request back to plugin, as a single object
-    return `<a class="noteTitle sectionItem" {()=>onClickDashboardItem({itemID: '${itemID}', type: 'showNoteInEditorFromFilename', encodedFilename: '${encodeURIComponent(dateFilename)}', encodedContent: ''}}><i class="fa-regular fa-file-lines pad-right"></i> ${NPDateStr}</a>`
-  }
-  catch (error) {
+    return `<a class="noteTitle sectionItem" {()=>onClickDashboardItem({itemID: '${itemID}', type: 'showNoteInEditorFromFilename', encodedFilename: '${encodeURIComponent(
+      dateFilename,
+    )}', encodedContent: ''}}><i class="fa-regular fa-file-lines pad-right"></i> ${NPDateStr}</a>`
+  } catch (error) {
     logError('makeNoteTitleWithOpenActionFromNPDateStr', `${error.message} for input '${NPDateStr}'`)
     return '(error)'
   }
@@ -533,7 +513,7 @@ export function extendParasToAddStartTime(paras: Array<TParagraph>): Array<any> 
         if (startTimeStr[1] === ':') {
           startTimeStr = `0${startTimeStr}`
         }
-        if (startTimeStr.endsWith("PM")) {
+        if (startTimeStr.endsWith('PM')) {
           startTimeStr = String(Number(startTimeStr.slice(0, 2)) + 12) + startTimeStr.slice(2, 5)
         }
         logDebug('extendParaToAddStartTime', `found timeStr: ${thisTimeStr} from timeblock ${thisTimeStr}`)
@@ -541,14 +521,13 @@ export function extendParasToAddStartTime(paras: Array<TParagraph>): Array<any> 
         extendedPara.timeStr = startTimeStr
       } else {
         // $FlowIgnore(prop-missing)
-        extendedPara.timeStr = "none"
+        extendedPara.timeStr = 'none'
       }
       extendedParas.push(extendedPara)
     }
 
     return extendedParas
-  }
-  catch (error) {
+  } catch (error) {
     logError('dashboard / extendParaToAddTimeBlock', `${JSP(error)}`)
     return []
   }
@@ -567,21 +546,20 @@ export function extendParasToAddStartTime(paras: Array<TParagraph>): Array<any> 
 export function getStartTimeFromPara(para: TParagraph): any {
   try {
     // logDebug('getStartTimeFromPara', `starting with ${String(paras.length)} paras`)
-    let startTimeStr = "none"
+    let startTimeStr = 'none'
     const thisTimeStr = getTimeBlockString(para.content)
     if (thisTimeStr !== '') {
       startTimeStr = thisTimeStr.split('-')[0]
       if (startTimeStr[1] === ':') {
         startTimeStr = `0${startTimeStr}`
       }
-      if (startTimeStr.endsWith("PM")) {
+      if (startTimeStr.endsWith('PM')) {
         startTimeStr = String(Number(startTimeStr.slice(0, 2)) + 12) + startTimeStr.slice(2, 5)
       }
       logDebug('extendParaToAddStartTime', `found timeStr: ${startTimeStr} from timeblock ${thisTimeStr}`)
     }
     return startTimeStr
-  }
-  catch (error) {
+  } catch (error) {
     logError('dashboard / extendParaToAddTimeBlock', `${JSP(error)}`)
     return []
   }
@@ -600,7 +578,7 @@ export function getStartTimeFromPara(para: TParagraph): any {
  */
 export function makeFakeCallbackButton(buttonText: string, pluginName: string, commandName: string, commandArgs: string, tooltipText: string = ''): string {
   const xcallbackURL = createRunPluginCallbackUrl(pluginName, commandName, commandArgs)
-  const output = (tooltipText)
+  const output = tooltipText
     ? `<span class="fake-button tooltip"><a class="button" href="${xcallbackURL}">${buttonText}</a><span class="tooltiptext">${tooltipText}</span></span>`
     : `<span class="fake-button"><a class="button" href="${xcallbackURL}">${buttonText}</a></span>`
   return output
@@ -621,7 +599,7 @@ export function makeFakeCallbackButton(buttonText: string, pluginName: string, c
  */
 export function makeRealCallbackButton(buttonText: string, pluginName: string, commandName: string, commandArgs: string, tooltipText: string = ''): string {
   const xcallbackURL = createRunPluginCallbackUrl(pluginName, commandName, commandArgs)
-  const output = (tooltipText)
+  const output = tooltipText
     ? `<button class="XCBButton tooltip"><a href="${xcallbackURL}">${buttonText}</a><span class="tooltiptext">${tooltipText}</span></button>`
     : `<button class="XCBButton"><a href="${xcallbackURL}">${buttonText}</a></button>`
   return output
@@ -673,13 +651,16 @@ export async function moveItemBetweenCalendarNotes(NPFromDateStr: string, NPToDa
       // so replace with one half of /qath:
       const shouldAppend = await getSettingFromAnotherPlugin('jgclark.QuickCapture', 'shouldAppend', false)
       const matchedHeading = findHeadingStartsWith(toNote, headingToPlaceUnder)
-      logDebug('addTextToNoteHeading', `Adding line '${targetContent}' to '${displayTitleWithRelDate(toNote)}' below matchedHeading '${matchedHeading}' (heading was '${headingToPlaceUnder}')`)
+      logDebug(
+        'addTextToNoteHeading',
+        `Adding line '${targetContent}' to '${displayTitleWithRelDate(toNote)}' below matchedHeading '${matchedHeading}' (heading was '${headingToPlaceUnder}')`,
+      )
       if (matchedHeading !== '') {
         // Heading does exist in note already
         toNote.addParagraphBelowHeadingTitle(
           targetContent,
           itemType,
-          (matchedHeading !== '') ? matchedHeading : headingToPlaceUnder,
+          matchedHeading !== '' ? matchedHeading : headingToPlaceUnder,
           shouldAppend, // NB: since 0.12 treated as position for all notes, not just inbox
           true, // create heading if needed (possible if supplied via headingArg)
         )
@@ -687,9 +668,7 @@ export async function moveItemBetweenCalendarNotes(NPFromDateStr: string, NPToDa
         const headingLevel = await getSettingFromAnotherPlugin('jgclark.QuickCapture', 'headingLevel', 2)
         const headingMarkers = '#'.repeat(headingLevel)
         const headingToUse = `${headingMarkers} ${headingToPlaceUnder}`
-        const insertionIndex = shouldAppend
-          ? findEndOfActivePartOfNote(toNote) + 1
-          : findStartOfActivePartOfNote(toNote)
+        const insertionIndex = shouldAppend ? findEndOfActivePartOfNote(toNote) + 1 : findStartOfActivePartOfNote(toNote)
         logDebug('moveItemBetweenCalendarNotes', `- adding new heading '${headingToUse}' at line index ${insertionIndex} ${shouldAppend ? 'at end' : 'at start'}`)
         toNote.insertParagraph(headingToUse, insertionIndex, 'text') // can't use 'title' type as it doesn't allow headingLevel to be set
         logDebug('moveItemBetweenCalendarNotes', `- then adding text '${targetContent}' after `)